backend/requirements.txt
absolufy-imports==0.3.1
acres==0.5.0
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.2
docstring_parser==0.17.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.119.0
filelock==3.19.1
fitz==0.0.1.dev2
frontend==0.0.3
google-adk==1.10.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.120.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.3
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.25.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.44.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.2
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-resourcedetector-gcp==1.10.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
proto-plus==1.26.1
protobuf==6.32.1
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydot==4.0.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
requests==2.32.5
rsa==4.9.1
scipy==1.13.1
shapely==2.0.7
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
tools==1.0.6
traits==7.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
